+++ src/client/hooks/useDefaultRoute.ts
@@ -1,31 +1,40 @@
-import { matchPath, useHistory, useLocation } from 'react-router';
-import { routes } from 'virtual:routes';
-import sidebars from 'virtual:sidebars';
-import { isEmpty } from 'lodash-es';
+import { matchPath, useLocation, useMatch, useNavigate } from "react-router";
+import { routes } from "virtual:routes";
+import sidebars from "virtual:sidebars";
+import { isEmpty } from "lodash-es";
+import { useEffect } from "react";
+
+const findFirstRoute = () => {
+  const find = (data: SidebarNode[]) => {
+    for (let i = 0; i < data.length; i++) {
+      const curr = data[i];
+      if (isEmpty(curr.children)) {
+        return curr?.path;
+      } else {
+        return curr?.children?.[0]?.path;
+      }
+    }
+  };
+
+  return find(sidebars);
+};
 
 export const useDefaultRoute = () => {
-  const findFirstRoute = () => {
-    const find = (data: SidebarNode[]) => {
-      for (let i = 0; i < data.length; i++) {
-        const curr = data[i];
-        if (isEmpty(curr.children)) {
-          return curr?.path;
-        } else {
-          return curr?.children?.[0]?.path;
-        }
+  const match = useMatch("/");
+  const navigate = useNavigate();
+
+  useEffect(() => {
+    if (match) {
+      if (match.pathname === "/") {
+        navigate(
+          {
+            pathname: findFirstRoute() || "/index",
+          },
+          {
+            replace: true,
+          }
+        );
       }
-    };
-
-    return find(sidebars);
-  };
-  const history = useHistory();
-  const location = useLocation();
-  const result = matchPath(
-    location.pathname,
-    routes.map((o) => o.path)
-  );
-
-  if (result.path === '*' && location.pathname === '/') {
-    history.push(findFirstRoute() || '/index');
-  }
+    }
+  }, [match]);
 };