+++ src/client/index.tsx
@@ -1,7 +1,6 @@
 import React, { useEffect, Suspense } from "react";
 import ReactDOM from "react-dom";
-import { HashRouter, Switch, Route } from "react-router-dom";
-import { useLocation } from "react-router";
+import { HashRouter, Route, Routes } from "react-router-dom";
 import { Document } from "./components/Document";
 import { Header } from "./components/Header";
 import { Sidebar } from "./components/Sidebar";
@@ -11,53 +10,61 @@ import { GlobalStyle, CssVariables } from "./styled";
 import Provider from "virtual:provider";
 import appData from "virtual:appData";
 import { routes } from "virtual:routes";
+import { useDefaultRoute } from "./hooks/useDefaultRoute";
 
-function RouteWithSubRoutes(route) {
-  return (
-    <Route
-      path={route.path}
-      render={(props) => (
-        <Suspense fallback={<></>}>
-          <route.component {...props} routes={route.routes} />
-        </Suspense>
-      )}
-    />
-  );
-}
-
-const App = () => {
-  const location = useLocation();
+const App: React.FC = ({ children }) => {
   useEffect(() => {
     document.title = appData.title;
     console.log(`%cDocit@${appData.version}`, `color: #9B1D30;`);
   }, []);
 
   return (
+    <>
+      <Header />
+      <Sidebar />
+      <Document>{children}</Document>
+    </>
+  );
+};
+
+const Root = () => {
+  useDefaultRoute();
+
+  return (
     <Provider>
       <CssVariables />
       <GlobalStyle />
-      {location.pathname !== "/sandbox" ? (
-        <>
-          <Header />
-          <Sidebar />
-          <Document />
-        </>
-      ) : (
-        <Switch>
-          {routes
-            .filter((o) => o.path === "/sandbox")
-            .map((route) => (
-              <RouteWithSubRoutes key={route.path} {...route} />
-            ))}
-        </Switch>
-      )}
+      <Routes>
+        {routes.map((route) => {
+          const Component = route.component;
+          return (
+            <Route
+              key={route.path}
+              path={route.path}
+              element={
+                route.path !== "/__sandbox__" ? (
+                  <App>
+                    <Suspense fallback={<></>}>
+                      <Component />
+                    </Suspense>
+                  </App>
+                ) : (
+                  <Suspense fallback={<></>}>
+                    <Component />
+                  </Suspense>
+                )
+              }
+            />
+          );
+        })}
+      </Routes>
     </Provider>
   );
 };
 
 ReactDOM.render(
   <HashRouter>
-    <App />
+    <Root />
   </HashRouter>,
   document.getElementById("app")
 );