+++ CHANGELOG.md
@@ -14,6 +14,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## [0.10.0] - Unreleased
+
+### Changed
+
+- Change StyledComponents naming
+- Change enum styling from color red to font-weight bold
+- Update react-router to v6
+
 ## [0.9.1] - 2022-07-06
 
 ### Changed